Restore uint64 unique_id and original pc_generate_unique_id formula

+++ src/common/mmo.hpp
@@ -262,7 +262,7 @@ struct item {
 	t_itemid card[MAX_SLOTS];
 	struct s_item_randomoption option[MAX_ITEM_RDM_OPT];		// max of 5 random options can be supported.
 	unsigned int expire_time;
-	uint32 unique_id;
+	uint64 unique_id;
 	unsigned int equipSwitch; // location(s) where item is equipped for equip switching (using enum equip_pos for bitmasking)
 };
